requirements.txt
asgiref==3.3.4
astroid==2.5.6
boto3==1.17.78
botocore==1.20.78
colorama==0.4.4
Django==3.2.3
django-filter==2.4.0
django-storages==1.11.1
gunicorn==20.1.0
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
numpy==1.20.3
pandas==1.2.4